Moving stage summary list item.
