+++ app/src/main/java/com/adammcneilly/pocketleague/ui/EventDetailContent.kt
@@ -33,6 +33,7 @@ import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.adammcneilly.pocketleague.composables.components.Chip
+import com.adammcneilly.pocketleague.composables.eventstage.StageSummaryListItem
 import com.adammcneilly.pocketleague.composables.team.TeamOverviewListItem
 import com.adammcneilly.pocketleague.core.displaymodels.EventDetailDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.EventStageSummaryDisplayModel
